+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [2.3.3]
 
+### Added 
+
+- Adding support for secure and unsecure configurations in pipeline simulation - https://github.com/DynamicsValue/fake-xrm-easy/issues/103
+
 ### Changed
 
 - Update legacy CRM SDK 2011 dependency to use official MS package - https://github.com/DynamicsValue/fake-xrm-easy/issues/105